cleaned up tests for HomeFragment
