+++ moodBookProject/app/src/androidTest/java/com/example/moodbook/HomeFragmentTest.java
@@ -39,7 +39,7 @@ public class HomeFragmentTest {
 
 
     @Before
-    public void setUp() throws Exception{
+    public void setUp() {
         solo = new Solo(InstrumentationRegistry.getInstrumentation(), rule.getActivity());
         TestHelper.setup(solo);
     }
@@ -49,7 +49,7 @@ public class HomeFragmentTest {
      * Check if moods are sorted starting from most recent
      */
     @Test
-    public void checkSorting(){
+    public void testSorting(){
         // wait for activity to change
         solo.sleep(5000);
         solo.assertCurrentActivity("Wrong Activity", MainActivity.class);
@@ -85,8 +85,8 @@ public class HomeFragmentTest {
     /**
      * Check if filter filters out mood with non-matching emotion
      */
-    //@Test
-    public void checkFiltering() {
+    @Test
+    public void testFiltering() {
         // wait for activity to change
         solo.sleep(5000);
         solo.assertCurrentActivity("Wrong Activity", MainActivity.class);
