+++ moodBookProject/app/src/androidTest/java/com/example/moodbook/UpdateMoodTest.java
@@ -25,28 +25,7 @@ public class UpdateMoodObjectTest {
     @Before
     public void setUp() throws Exception{
         solo = new Solo(InstrumentationRegistry.getInstrumentation(), rule.getActivity());
-        // logout if logged in
-        if (solo.searchText("Mood History")){
-            solo.clickOnImageButton(0);
-            for (int i = 0; i < 4; i++){
-                solo.sendKey(Solo.DOWN);
-            }
-            solo.clickOnText("Logout");
-            solo.sleep(3000);
-        }
-        // login with test account
-        if (solo.searchText("login")){
-            login();
-        }
-    }
-
-    /**
-     * used in tests to first login to the app
-     */
-    public void login(){
-        solo.enterText((EditText) solo.getView(R.id.email), "test@test.com");
-        solo.enterText((EditText) solo.getView(R.id.password), "testtest");
-        solo.clickOnButton("login");
+        TestHelper.setup(solo);
     }
