--- moodBookProject/app/src/test/java/com/example/moodbook/AddMoodTest.java
@@ -1,17 +1,0 @@
-package com.example.moodbook;
-import com.example.moodbook.ui.login.DBAuth;
-import com.google.firebase.auth.FirebaseAuth;
-import com.google.firebase.auth.FirebaseUser;
-import com.google.firebase.firestore.FirebaseFirestore;
-
-import org.junit.Test;
-import org.junit.runner.RunWith;
-import org.mockito.Mock;
-import org.mockito.Mockito;
-import org.mockito.runners.MockitoJUnitRunner;
-
-import static org.junit.jupiter.api.Assertions.*;
-@RunWith(MockitoJUnitRunner.class)
-public class AddMoodTest {
-
-}
